requirements.txt
jax
jaxlib
viser
numpy
scipy
pillow
tqdm
optax
chex
matplotlibjax-metal
